Improve plugin API to support flexible configuration
Description:
- Add handling of `self.configure` method that is called when
plugin is used. I.e. `plugin.used` methond is called one time
on first usage, and `plugin.configure` is called every time when
`use :some_plugin` is called. This allows to extend configuration
of already included extensions. For example, `rescue` plugin depends
on `:status` plugin and uses it with no arguments. After that, it
should be still possible to `use :status` plugin explicitly in
service itself with necessary arguments.